Added login and user management system

+++ part13/controllers/blogs.js
@@ -1,14 +1,31 @@
+require('dotenv').config()
 const blogsRouter = require('express').Router()
-const { Blog } = require('../models')
+const { User, Blog } = require('../models')
+const { tokenExtractor } = require('../util/middleware')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (req, res) => {
-  const blogs = await Blog.findAll()
+  const blogs = await Blog.findAll({
+    attributes: { exclude: ['userId'] },
+    include: {
+      model: User,
+      attributes: ['name']
+    }
+  })
   res.json(blogs)
 })
   
-blogsRouter.post('/', async (req, res) => {
+blogsRouter.post('/', tokenExtractor, async (req, res) => {
   try {
-    const blog = await Blog.create(req.body)
+    const decodedToken = jwt.verify(req.token, process.env.SECRET)
+    if (!decodedToken) {
+      return res.status(401).json({
+        error: 'invalid token'
+      })
+    }
+
+    const user = await User.findByPk(decodedToken.id)
+    const blog = await Blog.create({ ...req.body, userId: user.id, date: new Date() })
     return res.json(blog)
   } catch (error) {
     return res.status(400).json({ error })
@@ -24,14 +41,26 @@ blogsRouter.get('/:id', async (req, res) => {
   }
 })
   
-blogsRouter.delete('/:id', async (req, res) => {
+blogsRouter.delete('/:id', tokenExtractor, async (req, res) => {
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+  if (!decodedToken) {
+    return res.status(401).json({
+      error: 'invalid token'
+    })
+  }
   const blog = await Blog.findByPk(req.params.id)
-  await Blog.destroy({
-    where: {
-      id: blog.id
-    } 
-  })
-  return res.status(204).end()
+  if (blog.userId === decodedToken.id) {
+    await Blog.destroy({
+      where: {
+        id: blog.id
+      } 
+    })
+    return res.status(204).end()
+  } else {
+    return res.status(401).json({
+      error: 'invalid user'
+    })
+  }
 })
 
 blogsRouter.put('/:id', async (req, res) => {
